src/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
PyMuPDF==1.23.6
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.0.0
exa-py
slowapi==0.1.8
prometheus-fastapi-instrumentator
psutil==5.9.5
